+++ src/main/java/com/becareful/becarefulserver/domain/socialworker/controller/ElderlyController.java
@@ -42,7 +42,7 @@ public class ElderlyController {
     }
 
     //어르신 목록 - 성함, 나이, 성별, 프로필, 요양등급, 요양보호자 수, 매칭 중인지(공고 진행중인거), 검색어 입력
-    @Operation(summary = "어르신 목록 조회", description = "검색어를 입력하면 해당 이름을 포함한 어르신 목록을 반환합니다.")
+    @Operation(summary = "어르신 검색", description = "검색어를 입력하면 해당 이름을 포함한 어르신 목록을 반환합니다.")
     @GetMapping("/search")
     public ResponseEntity<List<ElderlyListResponse>> getElderlyListBySearch(
             @Parameter(name = "searchString", description = "검색어 (어르신 이름 일부 또는 전체)", example = "홍길동")
